api/requirements.txt
facenet-pytorch==2.5.3
Flask==2.3.2
gunicorn==21.2.0
joblib==1.3.1
# opencv-python==4.8.0.74
pillow>=10.0.1
requests==2.31.0
torch
Werkzeug==3.0.1
paho-mqtt